# Update dependencies