--- a/commons/src/main/java/com/powsybl/commons/compress/ZipPackager.java
+++ b/commons/src/main/java/com/powsybl/commons/compress/ZipPackager.java
@@ -10,8 +10,8 @@ package com.powsybl.commons.compress;
 import org.apache.commons.compress.archivers.zip.ZipArchiveEntry;
 import org.apache.commons.compress.archivers.zip.ZipArchiveOutputStream;
 import org.apache.commons.io.IOUtils;
+import org.jspecify.annotations.Nullable;
 
-import javax.annotation.Nullable;
 import java.io.*;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;