--- a/computation/src/main/java/com/powsybl/computation/ComputationExceptionBuilder.java
+++ b/computation/src/main/java/com/powsybl/computation/ComputationExceptionBuilder.java
@@ -7,10 +7,10 @@
  */
 package com.powsybl.computation;
 
+import org.jspecify.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.annotation.Nullable;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;