--- a/iidm/iidm-api/src/main/java/com/powsybl/iidm/network/Importer.java
+++ b/iidm/iidm-api/src/main/java/com/powsybl/iidm/network/Importer.java
@@ -14,8 +14,8 @@ import com.powsybl.commons.report.ReportNode;
 import com.powsybl.computation.ComputationManager;
 import com.powsybl.computation.local.LocalComputationManager;
 import com.powsybl.commons.parameters.Parameter;
+import org.jspecify.annotations.Nullable;
 
-import javax.annotation.Nullable;
 import java.util.*;
 import java.util.stream.Collectors;