--- a/pom.xml
+++ b/pom.xml
@@ -97,16 +97,16 @@
             ../../../distribution-core/target/site/jacoco-aggregate/jacoco.xml
         </sonar.coverage.jacoco.xmlReportPaths>
 
-        <assertj.version>3.26.0</assertj.version>
+        <assertj.version>3.27.3</assertj.version>
         <autoservice.version>1.1.1</autoservice.version>
-        <commonscli.version>1.8.0</commonscli.version>
-        <commonscompress.version>1.26.2</commonscompress.version>
-        <commons-configuration2.version>2.10.1</commons-configuration2.version>
-        <commons-csv.version>1.11.0</commons-csv.version>
-        <commonsio.version>2.16.1</commonsio.version>
-        <commonslang3.version>3.14.0</commonslang3.version>
+        <commonscli.version>1.9.0</commonscli.version>
+        <commonscompress.version>1.27.1</commonscompress.version>
+        <commons-configuration2.version>2.11.0</commons-configuration2.version>
+        <commons-csv.version>1.13.0</commons-csv.version>
+        <commonsio.version>2.18.0</commonsio.version>
+        <commonslang3.version>3.17.0</commonslang3.version>
         <commonsmath3.version>3.6.1</commonsmath3.version>
-        <commonstext.version>1.12.0</commonstext.version>
+        <commonstext.version>1.13.0</commonstext.version>
         <ejml.version>0.43.1</ejml.version>
         <!-- findbugs cannot be updated until graphviz is updated/replaced since findbugs groupdId changed -->
         <findbugs-annotations.version>3.0.1</findbugs-annotations.version>
@@ -114,38 +114,37 @@
         <gdata.version>1.47.1_1</gdata.version>
         <!-- graphviz-builder seems abandoned - to be replaced ? -->
         <graphviz-builder.version>1.0.12</graphviz-builder.version>
-        <groovy.version>4.0.21</groovy.version>
-        <guava.version>33.2.1-jre</guava.version>
-        <jackson.version>2.17.1</jackson.version>
-        <jackson-databind.version>2.17.1</jackson-databind.version>
+        <groovy.version>4.0.26</groovy.version>
+        <guava.version>33.4.0-jre</guava.version>
+        <jackson.version>2.18.3</jackson.version>
         <jama.version>1.0.3</jama.version>
         <jgrapht.version>1.5.2</jgrapht.version>
         <jimfs.version>1.3.0</jimfs.version>
-        <junit-jupiter.version>5.10.2</junit-jupiter.version>
-        <logback.version>1.5.14</logback.version>
+        <junit-jupiter.version>5.12.0</junit-jupiter.version>
+        <logback.version>1.5.17</logback.version>
         <maven.compat>3.9.9</maven.compat>
-        <maven.core.version>3.8.5</maven.core.version>
-        <maven.plugin.annotations.version>3.13.0</maven.plugin.annotations.version>
+        <maven.core.version>3.9.9</maven.core.version>
+        <maven.plugin.annotations.version>3.15.1</maven.plugin.annotations.version>
         <maven.plugin.testing.harness.version>3.3.0</maven.plugin.testing.harness.version>
         <mfl.version>0.5.15</mfl.version>
-        <mockito.version>5.12.0</mockito.version>
+        <mockito.version>5.16.0</mockito.version>
         <nativelibloader.version>2.5.0</nativelibloader.version>
-        <poi.version>5.2.5</poi.version>
+        <poi.version>5.4.0</poi.version>
         <rdf4j.version>4.3.2</rdf4j.version>
         <sirocco.version>1.0</sirocco.version>
-        <slf4j.version>2.0.13</slf4j.version>
-        <snakeyaml.version>2.2</snakeyaml.version>
+        <slf4j.version>2.0.17</slf4j.version>
+        <snakeyaml.version>2.4</snakeyaml.version>
         <staxutils.version>20070216</staxutils.version>
         <stringtemplate.version>4.3.4</stringtemplate.version>
-        <testcontainers.version>1.19.8</testcontainers.version>
+        <testcontainers.version>1.20.6</testcontainers.version>
         <threeten.version>1.8.0</threeten.version>
         <!-- trove4j seems abandoned - to be replaced ? -->
         <trove4j.version>3.0.3</trove4j.version>
         <!-- univocity-parsers seems abandoned - to be replaced ? -->
         <univocity-parsers.version>2.9.0</univocity-parsers.version>
         <xmlunit.version>2.10.0</xmlunit.version>
-        <xz.version>1.9</xz.version>
-        <zstdjni.version>1.5.6-3</zstdjni.version>
+        <xz.version>1.10</xz.version>
+        <zstdjni.version>1.5.7-1</zstdjni.version>
 
         <powsyblmathnative.version>1.4.0</powsyblmathnative.version>
     </properties>
@@ -671,7 +670,7 @@
             <dependency>
                 <groupId>com.fasterxml.jackson.core</groupId>
                 <artifactId>jackson-databind</artifactId>
-                <version>${jackson-databind.version}</version>
+                <version>${jackson.version}</version>
             </dependency>
             <dependency>
                 <groupId>com.fasterxml.jackson.datatype</groupId>