**Please check if the PR fulfills these requirements**
<!-- please use `'[x]'` to check the checkboxes, or submit the PR and then click the checkboxes -->
- [x] The commit message follows our guidelines
- [x] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] A PR or issue has been opened in all impacted repositories (if any)


**Does this PR already have an issue describing the problem?**
No

**What kind of change does this PR introduce?**
Dependencies update

**What is the current behavior?**
Old dependencies

**What is the new behavior (if this is a feature change)?**

- New versions:

| ArtefactId | Old Version | New Version |
|------------|------------------|-------------------|
| assertj-core | 3.26.0 | 3.27.3 |
| commons-cli | 1.8.0 | 1.9.0 |
| commons-compress | 1.26.2 | 1.27.1 |
| commons-configuration2 | 2.10.1 | 2.12.0 |
| commons-csv | 1.11.0 | 1.14.0 |
| commons-io | 2.16.1 | 2.19.0 |
| commons-lang3 | 3.14.0 | 3.17.0 |
| commons-text | 1.12.0 | 1.13.1 |
| groovy | 4.0.21 | 4.0.27 |
| guava | 33.2.1-jre | 33.4.8-jre |
| jackson-core | 2.17.1 | 2.19.0 |
| jackson-databind | 2.17.1 | 2.19.0 |
| jackson-annotations | 2.17.1 | 2.19.0 |
| junit-jupiter | 5.10.2 | 5.12.2 |
| logback-classic | 1.5.14 | 1.5.18 |
| maven-core | 3.8.5 | 3.9.9 |
| maven-plugin-annotations | 3.13.0 | 3.15.1 |
| mockito-core | 5.12.0 | 5.18.0 |
| slf4j-api | 2.0.13 | 2.0.17 |
| snakeyaml | 2.2 | 2.4 |
| testcontainers | 1.19.8 | 1.21.0 |
| xmlunit-core | 2.10.0 | 2.10.2 |
| xz | 1.9 | 1.10 |
| zstd-jni | 1.5.6-3 | 1.5.7-3 |

- switch `@Nullable` annotation in code from findbugs-annotations to jspecify (via guava)

**Does this PR introduce a breaking change or deprecate an API?**
- [ ] Yes
- [x] No

**If yes, please check if the following requirements are fulfilled**
<!-- If no breaking changes or API deprecations were introduced, delete this section -->
- [ ] The *Breaking Change* or *Deprecated* label has been added
- [ ] The migration steps are described in the following section

**What changes might users need to make in their application due to this PR? (migration steps)**
<!-- If this PR introduces a breaking change, describe the migration steps -->
<!-- The content of this section will be copied in the migration guide -->



**Other information**:
Performance comparison on my computer:
| Module                                      | Before (s) | After (s) | Diff (s)  | Evolution (%) |
|---------------------------------------------|-----------|-----------|----------|--------------|
| powsybl                                     | 0.857     | 0.669     | -0.188   | -21.94%      |
| Commons test                                | 1.211     | 1.232     | +0.021   | +1.73%       |
| Commons                                     | 7.159     | 6.796     | -0.363   | -5.07%       |
| Computation API                             | 2.072     | 1.779     | -0.293   | -14.14%      |
| Tools                                       | 1.668     | 1.585     | -0.083   | -4.98%       |
| Tools test                                  | 1.296     | 1.313     | +0.017   | +1.31%       |
| Local computation                           | 4.655     | 4.570     | -0.085   | -1.83%       |
| Math                                        | 2.102     | 2.126     | +0.024   | +1.14%       |
| IIDM                                        | 0.018     | 0.021     | +0.003   | +16.67%      |
| IIDM network model                          | 4.638     | 5.358     | +0.720   | +15.53%      |
| IIDM common extensions                      | 0.911     | 1.200     | +0.289   | +31.74%      |
| Config Test                                 | 0.562     | 0.625     | +0.063   | +11.21%      |
| IIDM testing networks                       | 0.403     | 0.466     | +0.063   | +15.64%      |
| IIDM TCK                                    | 2.149     | 2.367     | +0.218   | +10.15%      |
| IIDM network model implementation           | 5.996     | 6.900     | +0.904   | +15.08%      |
| IIDM serializer/deserializer                | 15.674    | 16.736    | +1.062   | +6.77%       |
| Load-flow                                   | 0.022     | 0.019     | -0.003   | -13.64%      |
| Load-flow API                               | 2.171     | 2.035     | -0.136   | -6.27%       |
| Network modifications                       | 5.291     | 5.064     | -0.227   | -4.29%       |
| Action                                      | 1.463     | 1.675     | +0.212   | +14.49%      |
| IIDM network criteria                       | 2.475     | 2.596     | +0.121   | +4.89%       |
| Contingency                                 | 0.017     | 0.015     | -0.002   | -11.76%      |
| Contingency API                             | 2.619     | 2.482     | -0.137   | -5.23%       |
| Contingency DSL                             | 3.626     | 3.714     | +0.088   | +2.43%       |
| Scripting                                   | 4.587     | 4.616     | +0.029   | +0.63%       |
| IIDM scripting                              | 4.909     | 5.204     | +0.295   | +6.01%       |
| CGMES conversion                            | 56.409    | 56.265    | -0.144   | -0.26%       |
| Time series API                             | 3.748     | 4.051     | +0.303   | +8.09%       |
| Dynamic Simulation API                      | 1.716     | 1.687     | -0.029   | -1.69%       |
| Dynamic Security analysis                   | 2.368     | 2.279     | -0.089   | -3.76%       |
| IEEE CDF model                              | 0.968     | 1.078     | +0.110   | +11.37%      |
| MATPOWER converter                          | 2.633     | 2.596     | -0.037   | -1.41%       |
| PSSE converter                              | 4.308     | 4.443     | +0.135   | +3.13%       |
| PowerFactory model                          | 0.915     | 0.881     | -0.034   | -3.72%       |
| PowerFactory converter                      | 2.580     | 2.604     | +0.024   | +0.93%       |
| Security Analysis Default Implementation    | 2.359     | 2.035     | -0.324   | -13.73%      |
| Sensitivity analysis API                    | 2.343     | 2.341     | -0.002   | -0.09%       |
| Short-circuit API                           | 2.731     | 2.521     | -0.210   | -7.69%       |
| Time series DSL                             | 6.867     | 6.966     | +0.099   | +1.44%       |
| **Total**                                     | **247.00**    | **249.00**   | **+2.00**    | **+0.81%**       |